Complete tic-tac-toe
Adds necessary functions to tic-tac-tow.js to allow
game to work appropriately. Game will identify all
normal win conditions as well as a tie condition and
alert players to a win causing the board to clear. This
was pretty challenging given familiarity with JS and
jQuery. The most challenging part was identifying
the win and tie conditions. I'm okay with how it
came out, but there is probably a cleaner way to
get it done. Worked hard to break logic out into
separate functions to maintain SRP. Overall, very
happy with how it came out and with the JS/jQuery
exploration that was done.